# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopEZ API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  buyer: 'buyer',
  seller: 'seller',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  buyer: 'buyer',
  seller: 'seller',
} as const;

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  discountPercent: number;
  category: string;
  imageUrl: string;
  stock: number;
  sellerId: number;
  sellerName: string;
  rating: number;
  reviewCount: number;
  createdAt: string;
}

export interface ProductList {
  products: Product[];
  total: number;
  page: number;
  totalPages: number;
}

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  discountPercent: number;
  category: string;
  imageUrl: string;
  stock: number;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  discountPercent?: number;
  category?: string;
  imageUrl?: string;
  stock?: number;
}

export interface Category {
  name: string;
  count: number;
}

export interface Review {
  id: number;
  productId: number;
  userId: number;
  userName: string;
  rating: number;
  comment: string;
  createdAt: string;
}

export interface ReviewInput {
  rating: number;
  comment: string;
}

export interface CartItem {
  productId: number;
  quantity: number;
  product: Product;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  discount: number;
  total: number;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  price: number;
  imageUrl: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  buyerId: number;
  buyerName: string;
  items: OrderItem[];
  subtotal: number;
  discount: number;
  total: number;
  status: OrderStatus;
  shippingAddress: string;
  createdAt: string;
}

export interface OrderInput {
  shippingAddress: string;
  paymentMethod: string;
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  pending: 'pending',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface SellerStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  pendingOrders: number;
  avgOrderValue: number;
  revenueGrowth: number;
}

export interface SalesDataPoint {
  date: string;
  revenue: number;
  orders: number;
}

export interface TopProduct {
  productId: number;
  name: string;
  imageUrl: string;
  totalSold: number;
  revenue: number;
}

export type ListProductsParams = {
category?: string;
search?: string;
sort?: ListProductsSort;
minPrice?: number;
maxPrice?: number;
page?: number;
limit?: number;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  rating: 'rating',
  newest: 'newest',
} as const;

export type GetSellerSalesParams = {
period?: GetSellerSalesPeriod;
};

export type GetSellerSalesPeriod = typeof GetSellerSalesPeriod[keyof typeof GetSellerSalesPeriod];


export const GetSellerSalesPeriod = {
  '7d': '7d',
  '30d': '30d',
  '90d': '90d',
} as const;

export type ListSellerOrdersParams = {
status?: string;
};