# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopEZ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  shippingAddress: string;
  paymentMethod: string;
}