# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ShopEZ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  discountPercent: number;
  category: string;
  imageUrl: string;
  stock: number;
  sellerId: number;
  sellerName: string;
  rating: number;
  reviewCount: number;
  createdAt: string;
}